refactor of idf download and css classes
